[47] code cleanup
